% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_functions.R
\name{plot.shrinkTVPVAR}
\alias{plot.shrinkTVPVAR}
\title{Plotting method for \code{shrinkTVPVAR} objects}
\usage{
\method{plot}{shrinkTVPVAR}(x, ...)
}
\arguments{
\item{x}{\code{shrinkTVPVAR} object}

\item{...}{further arguments to be passed to \code{\link{state_plotter}} function.

For further details see \code{\link{state_plotter}} function.}
}
\value{
Called for its side effects and returns invisibly.
}
\description{
Plotting method for \code{shrinkTVPVAR} objects
}
\examples{
\donttest{
set.seed(123)
sim <- simTVPVAR(p = 2)
data <- sim$data

res <- shrinkTVPVAR(data, p = 2)

plot(res)
}
}
\seealso{
Other plotting functions: 
\code{\link{TV_heatmap}()},
\code{\link{density_plotter}()},
\code{\link{plot.mcmc.tvp.var}()},
\code{\link{plot.mcmc.var}()},
\code{\link{plot.shrinkTVPVAR_fit}()},
\code{\link{plot.shrinkTVPVAR_forc}()},
\code{\link{state_plotter}()}
}
\author{
Peter Knaus \email{peter.knaus@wu.ac.at}
}
\concept{plotting functions}
