% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrinkTVPVAR_methods.R
\name{print.shrinkDTVPVAR}
\alias{print.shrinkDTVPVAR}
\title{Nicer printing of shrinkDTVPVAR objects}
\usage{
\method{print}{shrinkDTVPVAR}(x, ...)
}
\arguments{
\item{x}{a \code{shrinkDTVPVAR} object.}

\item{...}{Currently ignored.}
}
\value{
Called for its side effects and returns invisibly.
}
\description{
Nicer printing of shrinkDTVPVAR objects
}
\author{
Peter Knaus \email{peter.knaus@wu.ac.at}
}
