% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shrinkTVPVAR_methods.R
\name{print.shrinkTVPVAR}
\alias{print.shrinkTVPVAR}
\title{Nicer printing of shrinkTVPVAR objects}
\usage{
\method{print}{shrinkTVPVAR}(x, ...)
}
\arguments{
\item{x}{a \code{shrinkTVPVAR} object.}

\item{...}{Currently ignored.}
}
\value{
Called for its side effects and returns invisibly.
}
\description{
Nicer printing of shrinkTVPVAR objects
}
\author{
Peter Knaus \email{peter.knaus@wu.ac.at}
}
